A few more grass pieces makes things more interesting.
Possibly I should fill-in all the road forking and dead-ending types on
both grass and dirt.

Maybe there should be a way to make carctiles filled with a bunch of
"undefined"s and then be able to stamp a carctile onto another one to
create a combined one... oh i guess that's just something i want now
but won't be useful long-term.