Jesus Fucking Christ, I can't figure this out.
I'm trying to put an image on a button, and I'd like to do it
in a way that doesn't require me to know that the image will
be on the button ahead of time.

I've tried it 4 different ways, and I have the image specified
as a kind of background too, to try to just hardcode the fucker in.
Nothing works, and my whole afternoon yesterday was frustrating
fuckering so I can't do it right now. I can circle back.