Reintroduce the 'paths' Component to fix-up walk animations.
The animation trigger was just whether a path existed, but it
didn't know whether the path was finished. Perhaps I can change
things up so that we take the "server" path data and copy it to
some client records that can do things different... but I just
went back to storing the path information in a separate component
that I keep updated as the path is followed. Once you finish the path,
it's removed.