Change the model's randomPositionFrom().
The pathfinding library just looks at every face's centroid, checks
the distanceSquared (contradicting documentation) and if it passes
it adds it to a list of candidates from which it randomly picks.

That resulted in a lot of monsters standing in the same spots.

My new algorithm is pretty much the same except the candidates are
random points on each face. Actual distance is used.