New button images from Jon.
So there are Down/Hover versions as well, but... I guess I'd set those
up as backgrounds. I already size the element, so maybe I just put a
spacer in the button.

Also would be nice to have a better way to specify "no background"
or "no border". Right now they are drawn, but all with a transparent
color.