Hack out some blippiness for now by hiding newly-loaded actors.
When the SceneSystem gets the callback that a GLTF has loaded, it
immediately replaces the placeholder object with the new Object3D,
but other systems and components don't get a chance to update it
until the next domain-wide update().

I'm actually still seeing some blips but it's much better now.