Copy three-pathfinding, make it load THREE via import.
This will allow us to fix some other wonky business and not have to do annoying
"put THREE in global scope" bullshit on the server, which was about to become
needed.

This doesn't yet remove the three-pathfinding dependency because I have other
stuff going on in package.json that I don't feel like isolating right now. Yay
for cowboy development.