Sorted out the walking condition nicely.
When you pass a Value as an input, the graph system actually creates a
node for it on-the-fly. If we go ahead and create that node, we can name
it and then we can update the Value in it later. Extract and map() a
value from the PathComponent, which we can only get after we've added
the entity and configured the graph!

I discovered this by checking the documentation, which right now means
asking Andrzej.