Finish changing to the Behavior system, new actor states.
So basically a Behavior can have its own stateful data about an actor.
That way, the actor's action/state can be independant of what the actor
is doing behind the scenes. Previously I needed 3 different walking
states but now there are 0.

For example the EatFoodBehavior has the actor go through various states
like Hungry, Eating, Sleepy and then it hands things off to the
SleepBehavior.

We could still separate the "emojional state" from state, but as of now
it's not necessary.

There is also still the "instant" which is sorta like a secondary state.
Perhaps we just send down an array of states...

NOTE: something is broken with sleep animation, I'll figure that out
soon.