Re-enable shadows from the primary light.
Shadows look so good. They are still disabled on android.

I tried adding shadows from the rimlight as well and that was cool, but
it looked most wrong in the "Day" scheme with a dimmer long shadow
aside the main shadow. So I nixed it.

The sunset shadows are so dope that I want to wire-up the light cycle
right now, but I'll try to stay focused on behaviors...