Created a LerpSystem that will lerp arbitrary properties on Object3Ds.
Configuring things is way smoother, because you specify the lerp
configuration in the entity config. So that hackery is removed.

But, maybe I worry about the inefficiency of looking up properties by
name, etc. Also there's a bit of repetition in the entity configs for
the 3 lights, where before I deferred to a method that avoided the
repeating. I guess I could make a method to populate the entity
configs...

I could move this somewhere like tfw/scene3/entity. I needed to expose
createObject3D() in order to utilize the Object3DConfig to sensibly load
the "sources" of lerping.