Give devs their own private Idaho.
Now that ranches are persistent again, we don't want everyone writing to the
same ranch record because it will be a nightmare to test things. So if you
don't specify a ranchid (on localhost), we make up a new one and save it to
browser local storage and then keep redirecting you to that one.

But you should retrain your browsers to use your personal ranch link because we
can't magically redirect you to the right ranch if you're using multiple
browsers. Maybe in a few days we'll just start requiring a ranchid, so that if
you forget, you just see a blank page instead of ending up at a randomly
generated new ranch.