Remove the small addition I made to doc.
Because I'm absolutely bonkers and it was doing the right thing.
Sure, it's calling distanceToSquared, but then it compares it to
nearRange * nearRange!

I spotted this while looking at the source for the Babylon navigation
plugin, which is basically this same code AGAIN.

Which isn't super inspiring. The maxDistanceSquared could be computed
once, outside the loop. If Babylon is supposedly better it'd be cool if
it didn't copy braindead code from the Three world.