Avatars!
Non-guests can click anywhere in the ranch to move their avatar their.
If they don't have an avatar yet: it is magic'd up.
When you disconnect your avatar is "frozen".

This highlights a problem with how I track time for paths: my solution
last week to doing a partial tick was fine if the avatar is stopping
mid-tick, but if it is continuing on its path then things get all
messed-up because the actor will participate in the next full tick.
Once it has done that, it has "seen" more time than has actually passed.
I need to track actual timestamps in some more places.