Attempt to fix my walking problem.
Well I'm clever enough to think of a hacky solution (thank the porcelin
muse!) but apparently not clever enough to make it work.

After starting a walk mid-tick, I can make things right for the next
full tick by altering the path segment so that the starting point is
projected back to where it would be if the path was started on the
*last tick*.

Should work, but something seems to be wrong. It's too cold today.