Have avatars do a 2x animation speed on their walk animation.
Since they walk 2x as fast as other actors.

I'll play with updating this after setup. When monsters rush for
food, we should be able to make that particular walk faster...
But right now I don't think there's a way with the AnimationController
in our graph system to update the timeScale while a particular animation
is playing.

I'm messing with this because Daniel wants me to explore "chat snake"
implementation.