Another fix for pathfinding.
AStar returns an empty array on failure. Let's check for failure!
Maybe this was intentional on the part of the Pathfinding authors.

The problem seems to happen when leaving the "dead zone" area, so maybe
this will all clear-up if that's fixed, but in the meanwhile let's
not return paths that leave the navmesh.