Trim the tail to the right length when creating.
Debug the length as we're adjusting it.
Grumble... it'd be more accurate to run through all the points rather
than measure the length delta to the new head and then remove that delta
from the end. I suppose the errors are tiny but who knows how long a
chat snake could be around?

So none of this accounts for smoothly updating the snake position on the
client, like we do for animating walks. Bleah.