Update the snake head on the client as the owner walks.
Publish the circleId of an actor, and have a new step that extends the
head of a snake to the current client position while pathing.

What a mess.

The intention was to work-out how chat snakes work and act, but I'm
currently saddled with the insanity of moncher paths and all that. At
the very least I should rejigger how the paths work but I wasn't going
to invest more time in that if we had this theoretical "space" system
for moving bodies around...